requirements.txt
asgiref==3.8.1
charset-normalizer==3.4.2
dj-database-url==3.0.1
Django==5.2
django-debug-toolbar==5.2.0
django-widget-tweaks==1.5.0
pillow==11.3.0
reportlab==4.4.2
sqlparse==0.5.3
tzdata==2025.2
whitenoise==6.9.0
